Fix the handling EXTENDED_ARG 0
The issue identified/fixed in PR 24 results from the fact that CPython sometimes emits meaningless EXTENDED_ARG (meaningless in that the argument fits in 8 bits). As a consequence, the size inferred from the argument leads to the wrong offset (since we miss one opcode) and breaks jump targets. To be safe against this, we manually keep track of the number of seen extended args to compute the size. Ideally CPython should not emit such bytecode.