Fix the handling EXTENDED_ARG 0 (#28)
* Fix the handling EXTENDED_ARG 0

The issue identified/fixed in PR 24 results from the fact that CPython sometimes emits meaningless EXTENDED_ARG (meaningless in that the argument fits in 8 bits). As a consequence, the size inferred from the argument leads to the wrong offset (since we miss one opcode) and breaks jump targets. To be safe against this, we manually keep track of the number of seen extended args to compute the size. Ideally CPython should not emit such bytecode.

* Allow to have EXTENDED_ARG in Bytecode.

Previously it was not exactly forbidden, but the conversion to code would fail because of a broken stack_effect computation.

* Ensure that we properly round trip through ConcreteBytecode

When CPython emitted useless EXTENDED_ARG 0, we need to put them back as otherwise we would corrupt the jump offsets. As soon as we go through Bytecode, we are safe.