cfg: generate cleaner CFG in the presence of an exception table
We do not senf a TryEnd following a final instruction to a new block, we insert TryEnd at the beginning of blocks which are a target of a jump that goes beyond the existing TryEnd in the bytecode. When converting back we keep only teh first seen TryEnd. This adds some constraints on what is allowed with TryBegin, TryEnd that should be documented.