fix: handle small try blocks
We make sure that we can handle small (and potentially empty) try
blocks. These are blocks where the end offset is smaller or at most
equal to the start offset. These blocks can occur when a single
instruction that requires EXTENDED_ARG opcodes is wrapped in a try
block. Original code objects seem to wrap the whole instruction,
including the EXTENDED_ARG opcodes. The fix in this PR wraps just
the instruction, leaving the extra EXTENDED_ARG opcodes outside the
try block. This might not be a problem in general, but it is perhaps
not ideal.